+++ AvatarGame/modello.cpp
@@ -116,35 +116,35 @@ void Modello::caricare()
                     bool spada = newAttributo.hasAttribute("Spada") ? newAttributo.value("Spada").toString() == "true" ? true : false : false;
                     bool anello = newAttributo.hasAttribute("Anello") ? newAttributo.value("Anello").toString() == "true" ? true : false : false;
                     bool libro = newAttributo.hasAttribute("Libro") ? newAttributo.value("Libro").toString() == "true" ? true : false : false;
-                    unsigned int trasparentia = newAttributo.hasAttribute("Trasparentia") ? newAttributo.value("Trasparentia").toUInt() : 0;
+                    double trasparentia = newAttributo.hasAttribute("Trasparentia") ? newAttributo.value("Trasparentia").toDouble() : 0;
                     inserire = new Elfo(nome, descrizione, lvl, exp, forza, magia, difesa, scienza, media, terreno, sesso, percorso, scudo, spada, anello, libro, trasparentia);
                 } else if(reader.name() == "Nano") {
                     bool scudo = newAttributo.hasAttribute("Scudo") ? newAttributo.value("Scudo").toString() == "true" ? true : false : false;
                     bool spada = newAttributo.hasAttribute("Spada") ? newAttributo.value("Spada").toString() == "true" ? true : false : false;
                     bool anello = newAttributo.hasAttribute("Anello") ? newAttributo.value("Anello").toString() == "true" ? true : false : false;
                     bool libro = newAttributo.hasAttribute("Libro") ? newAttributo.value("Libro").toString() == "true" ? true : false : false;
-                    unsigned int corteccia = newAttributo.hasAttribute("Corteccia") ? newAttributo.value("Corteccia").toUInt() : 0;
+                    double corteccia = newAttributo.hasAttribute("Corteccia") ? newAttributo.value("Corteccia").toDouble() : 0;
                     inserire = new Nano(nome, descrizione, lvl, exp, forza, magia, difesa, scienza, media, terreno, sesso, percorso, scudo, spada, anello, libro, corteccia);
                 } else if(reader.name() == "Umano") {
                     bool scudo = newAttributo.hasAttribute("Scudo") ? newAttributo.value("Scudo").toString() == "true" ? true : false : false;
                     bool spada = newAttributo.hasAttribute("Spada") ? newAttributo.value("Spada").toString() == "true" ? true : false : false;
                     bool anello = newAttributo.hasAttribute("Anello") ? newAttributo.value("Anello").toString() == "true" ? true : false : false;
                     bool libro = newAttributo.hasAttribute("Libro") ? newAttributo.value("Libro").toString() == "true" ? true : false : false;
-                    unsigned int ingegno = newAttributo.hasAttribute("Ingegnoscientifico") ? newAttributo.value("Ingegnoscientifico").toUInt() : 0; //qui invece avere lo spazio non conta nulla
+                    double ingegno = newAttributo.hasAttribute("Ingegnoscientifico") ? newAttributo.value("Ingegnoscientifico").toDouble() : 0; //qui invece avere lo spazio non conta nulla
                     inserire = new Umano(nome, descrizione, lvl, exp, forza, magia, difesa, scienza, media, terreno, sesso, percorso, scudo, spada, anello, libro, ingegno);
                 } else if(reader.name() == "Alieno") {
                     bool barriera = newAttributo.hasAttribute("Barriera") ? newAttributo.value("Barriera").toString() == "true" ? true : false : false;
                     bool laser = newAttributo.hasAttribute("Laser") ? newAttributo.value("Laser").toString() == "true" ? true : false : false;
                     bool amuleto = newAttributo.hasAttribute("Amuleto") ? newAttributo.value("Amuleto").toString() == "true" ? true : false : false;
                     bool chip = newAttributo.hasAttribute("Chip") ? newAttributo.value("Chip").toString() == "true" ? true : false : false;
-                    unsigned int ufo = newAttributo.hasAttribute("Ufo") ? newAttributo.value("Ufo").toUInt() : 0;
+                    double ufo = newAttributo.hasAttribute("Ufo") ? newAttributo.value("Ufo").toDouble() : 0;
                     inserire = new Alieno(nome, descrizione, lvl, exp, forza, magia, difesa, scienza, media, terreno, sesso, percorso, barriera, laser, amuleto, chip, ufo);
                 } else if(reader.name() == "Mostro") {
                     bool barriera = newAttributo.hasAttribute("Barriera") ? newAttributo.value("Barriera").toString() == "true" ? true : false : false;
                     bool laser = newAttributo.hasAttribute("Laser") ? newAttributo.value("Laser").toString() == "true" ? true : false : false;
                     bool amuleto = newAttributo.hasAttribute("Amuleto") ? newAttributo.value("Amuleto").toString() == "true" ? true : false : false;
                     bool chip = newAttributo.hasAttribute("Chip") ? newAttributo.value("Chip").toString() == "true" ? true : false : false;
-                    unsigned int porta = newAttributo.hasAttribute("Portademoniaca") ? newAttributo.value("Portademoniaca").toUInt() : 0;
+                    double porta = newAttributo.hasAttribute("Portademoniaca") ? newAttributo.value("Portademoniaca").toDouble() : 0;
                     inserire = new Mostro(nome, descrizione, lvl, exp, forza, magia, difesa, scienza, media, terreno, sesso, percorso, barriera, laser, amuleto, chip, porta);
                 }
